requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
enum34==1.1.10
iniconfig==1.1.1
jsonpath==0.82
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
pytest-order==1.0.1
pytest-pycharm==0.7.0
requests==2.28.1
urllib3==1.26.9
selenium
webdriver_manager
